+++ forui/lib/properties.dart
@@ -3,3 +3,4 @@ library forui.properties;
 export 'src/properties/focusable_properties.dart';
 export 'src/properties/form_field_properties.dart';
 export 'src/properties/labelled_properties.dart';
+export 'src/properties/popover_properties.dart';