+++ forui/lib/src/widgets/select_menu_tile.dart
@@ -20,7 +20,7 @@ import 'package:meta/meta.dart';
 /// * [FSelectMenuTileStyle] for customizing a select group's appearance.
 class FSelectMenuTile<T> extends FormField<Set<T>>
     with FTileMixin
-    implements FFormFieldProperties<Set<T>>, FFocusableProperties {
+    implements FFocusableProperties, FFormFieldProperties<Set<T>>, FPopoverProperties {
   /// The controller that controls the selected tiles.
   final FSelectGroupController<T> groupController;
 
@@ -53,28 +53,19 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
   /// The divider between select tiles. Defaults to [FTileDivider.indented].
   final FTileDivider divider;
 
-  /// The anchor of the menu to which the [tileAnchor] is aligned to.
-  ///
-  /// Defaults to [Alignment.bottomCenter] on Android and iOS, and [Alignment.topCenter] on all other platforms.
-  final Alignment menuAnchor;
+  @override
+  final Alignment popoverAnchor;
 
-  /// The anchor of the child to which the [menuAnchor] is aligned to.
-  ///
-  /// Defaults to [Alignment.topCenter] on Android and iOS, and [Alignment.bottomCenter] on all other platforms.
-  final Alignment tileAnchor;
+  @override
+  final Alignment childAnchor;
 
-  /// The shifting strategy used to shift a menu when it overflows out of the viewport. Defaults to
-  /// [FPortalFollowerShift.flip].
-  ///
-  /// See [FPortalFollowerShift] for more information on the different shifting strategies.
+  @override
   final Offset Function(Size, FPortalTarget, FPortalFollower) shift;
 
-  /// True if the popover is hidden when tapped outside of it. Defaults to true.
+  @override
   final bool hideOnTapOutside;
 
-  /// True if the follower should include the cross-axis padding of the anchor when aligning to it. Defaults to false.
-  ///
-  /// Diagonal corners are ignored.
+  @override
   final bool directionPadding;
 
   /// True if the menu should be automatically hidden after a menu option is selected. Defaults to false.
@@ -128,8 +119,8 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
     this.maxHeight = double.infinity,
     this.dragStartBehavior = DragStartBehavior.start,
     this.divider = FTileDivider.full,
-    this.menuAnchor = Alignment.topRight,
-    this.tileAnchor = Alignment.bottomRight,
+    this.popoverAnchor = Alignment.topRight,
+    this.childAnchor = Alignment.bottomRight,
     this.shift = FPortalFollowerShift.flip,
     this.hideOnTapOutside = true,
     this.directionPadding = false,
@@ -182,15 +173,15 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
               key: GlobalObjectKey(state._controller._popover),
               controller: state._controller._popover,
               style: menuStyle,
-              followerAnchor: menuAnchor,
-              targetAnchor: tileAnchor,
+              popoverAnchor: popoverAnchor,
+              childAnchor: childAnchor,
               shift: shift,
               hideOnTapOutside: hideOnTapOutside,
               directionPadding: directionPadding,
               autofocus: autofocus,
               focusNode: focusNode,
               onFocusChange: onFocusChange,
-              followerBuilder: (context, _, __) => ConstrainedBox(
+              popoverBuilder: (context, _, __) => ConstrainedBox(
                 constraints: BoxConstraints(maxWidth: menuStyle.maxWidth),
                 child: FSelectTileGroup<T>(
                   groupController: state._controller,
@@ -204,7 +195,7 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
                   children: menu,
                 ),
               ),
-              target: FTile(
+              child: FTile(
                 style: tileStyle,
                 prefixIcon: prefixIcon,
                 enabled: enabled,
@@ -259,8 +250,8 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
     this.maxHeight = double.infinity,
     this.dragStartBehavior = DragStartBehavior.start,
     this.divider = FTileDivider.full,
-    this.menuAnchor = Alignment.topRight,
-    this.tileAnchor = Alignment.bottomRight,
+    this.popoverAnchor = Alignment.topRight,
+    this.childAnchor = Alignment.bottomRight,
     this.shift = FPortalFollowerShift.flip,
     this.hideOnTapOutside = true,
     this.directionPadding = false,
@@ -314,15 +305,15 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
               key: GlobalObjectKey(state._controller._popover),
               controller: state._controller._popover,
               style: menuStyle,
-              followerAnchor: menuAnchor,
-              targetAnchor: tileAnchor,
+              popoverAnchor: popoverAnchor,
+              childAnchor: childAnchor,
               shift: shift,
               hideOnTapOutside: hideOnTapOutside,
               directionPadding: directionPadding,
               autofocus: autofocus,
               focusNode: focusNode,
               onFocusChange: onFocusChange,
-              followerBuilder: (context, _, __) => ConstrainedBox(
+              popoverBuilder: (context, _, __) => ConstrainedBox(
                 constraints: BoxConstraints(maxWidth: menuStyle.maxWidth),
                 child: FSelectTileGroup<T>.builder(
                   groupController: state._controller,
@@ -337,7 +328,7 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
                   count: count,
                 ),
               ),
-              target: FTile(
+              child: FTile(
                 style: tileStyle,
                 prefixIcon: prefixIcon,
                 enabled: enabled,
@@ -380,8 +371,8 @@ class FSelectMenuTile<T> extends FormField<Set<T>>
       ..add(DoubleProperty('maxHeight', maxHeight))
       ..add(EnumProperty('dragStartBehavior', dragStartBehavior))
       ..add(EnumProperty('divider', divider))
-      ..add(DiagnosticsProperty('menuAnchor', menuAnchor))
-      ..add(DiagnosticsProperty('tileAnchor', tileAnchor))
+      ..add(DiagnosticsProperty('popoverAnchor', popoverAnchor))
+      ..add(DiagnosticsProperty('childAnchor', childAnchor))
       ..add(ObjectFlagProperty.has('shift', shift))
       ..add(FlagProperty('hideOnTapOutside', value: hideOnTapOutside, ifTrue: 'hideOnTapOutside'))
       ..add(FlagProperty('directionPadding', value: directionPadding, ifTrue: 'directionPadding'))